django restframework/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiomysql==0.2.0
aiosignal==1.4.0
aiosmtplib==3.0.2
alembic==1.16.5
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
appwrite==13.0.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
DateTime==5.5
distro==1.9.0
Django==5.2
django-ckeditor-5==0.2.18
django-filter==25.1
django-tenants==3.9.0
djangorestframework==3.16.0
dnspython==2.8.0
docutils==0.21.2
email-validator==2.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
filetype==1.2.0
Flask==3.1.0
fonttools==4.58.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-groq==0.3.7
langchain-text-splitters==0.3.11
langsmith==0.4.27
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.5
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pefile==2023.2.7
pillow==11.2.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydicom==3.0.1
Pygments==2.19.1
pyinstall==0.1.4
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pymeshlab==2023.12.post3
PyMySQL==1.1.2
pynetdicom==3.0.4
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.47.3
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
upstash-redis==1.4.0
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zope.interface==7.2
zstandard==0.24.0
